requirements.txt
numpy
pandas
geoplot
geopandas
shapely
matplotlib
python-dotenv
